pyproject.toml
[project]
name = "tdd-task-manager-api"
version = "0.1.0"
description = ""
authors = [
    {name = "diego.velilla",email = "diegovelillarecio@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "bcrypt (==4.3.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]